#Docs : 컴포넌트분리와 useSearchParams 사용
